Create separate Benchmark for DefaultMeterObservationHandler
Focus the ObservationBenchmark on only benchmarking the Observation infrastructure.
Separately benchmark the DefaultMeterObservationHandler, an implementation of an ObservationHandler.
